Allow partial build extension to work with multiple sources
Update the partial build extension to support playbooks that contain
more than one source. Prior to this commit, only the branches or tags
of the first source would be updated. This could be problematic if the
starting path using for a specific branch had changed since the path
from the earlier source would now be used on an incorrect branch.

The updated code now filters sources based on the `refname` and only
includes those that match. Support for HEAD references can now only be
used if there is a single source.

If none of the sources match, then a full build is performed.